refactor type declarations to correct location
