+++ src/components/PluginTrends/Layout/PluginDetailModal.tsx
@@ -7,7 +7,7 @@ import PluginInstallationsPercentageChart from '../Charts/PluginInstallationsPer
 import PluginInstallationsPerVersion from '../Charts/PluginInstallationsPerVersionChart'
 import PluginInstallationsPercentagePerVersionChart from '../Charts/PluginInstallationsPercentagePerVersionPieChart'
 import { handleJSONDownload } from '../../../utils/jsonUtils'
-import { IPluginData } from '../../../data/plugins'
+import { IPluginData } from '../../../types/types'
 import CloseIcon from '@mui/icons-material/Close'
 
 interface PluginDetailModalProps {
